+++ Sources/ZcashLightClientKit/Repository/TransactionRepository.swift
@@ -19,6 +19,9 @@ protocol TransactionRepository {
     func findPendingTransactions(latestHeight: BlockHeight, offset: Int, limit: Int) async throws -> [ZcashTransaction.Overview]
     func findReceived(offset: Int, limit: Int) async throws -> [ZcashTransaction.Overview]
     func findSent(offset: Int, limit: Int) async throws -> [ZcashTransaction.Overview]
+    // sourcery: mockedName="findMemosForRawID"
+    func findMemos(for rawID: Data) async throws -> [Memo]
+    // sourcery: mockedName="findMemosForZcashTransaction"
     func findMemos(for transaction: ZcashTransaction.Overview) async throws -> [Memo]
     func getRecipients(for rawID: Data) async throws -> [TransactionRecipient]
     func getTransactionOutputs(for rawID: Data) async throws -> [ZcashTransaction.Output]